requirements-dev.txt
pytest==7.4.4
pytest-flask==1.3.0
coverage==7.4.1
flake8==7.0.0
black==24.3.0
isort==5.13.2
pre-commit==3.6.0
pytest
flake8
black